Refactor nagGroupItems2
