+++ demos/storybook/stories/drawer/with-multiple-DrawerNavGroups.tsx
@@ -5,26 +5,7 @@ import { boolean, text } from '@storybook/addon-knobs';
 import { StoryFnReactReturnType } from '@storybook/react/dist/client/preview/types';
 import React from 'react';
 import { DrawerStoryContext } from './util';
-import { navGroupItems1 } from './with-basic-config';
-
-export const navGroupItems2 = (context: DrawerStoryContext): NavItem[] => [
-    {
-        title: 'Overview',
-        itemID: 'group2-1',
-        icon: <Dashboard />,
-        onClick: (): void => {
-            context.store.set({ selected: 'group2-1' });
-        },
-    },
-    {
-        title: 'Timeline',
-        itemID: 'group2-2',
-        icon: <Toc />,
-        onClick: (): void => {
-            context.store.set({ selected: 'group2-2' });
-        },
-    },
-];
+import { navGroupItems1, navGroupItems2 } from './with-basic-config';
 
 export const withMultipleNavGroups = (context: DrawerStoryContext): StoryFnReactReturnType => (
     <Drawer open={boolean('open', true)}>
@@ -39,7 +20,7 @@ export const withMultipleNavGroups = (context: DrawerStoryContext): StoryFnReact
             <DrawerNavGroup
                 title={text('navGroup[1].title', 'Second DrawerNavGroup')}
                 activeItem={context.state.selected}
-                items={navGroupItems2(context)}
+                items={navGroupItems2}
             />
         </DrawerBody>
     </Drawer>
